inline sticky header classes
