+++ src/org/thoughtcrime/securesms/ContactSelectionListFragment.java
@@ -18,29 +18,32 @@ package org.thoughtcrime.securesms;
 
 
 import android.database.Cursor;
+import android.graphics.Canvas;
+import android.graphics.Rect;
 import android.os.Build;
+import android.os.Build.VERSION;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.LoaderManager;
 import android.support.v4.content.Loader;
+import android.support.v4.view.ViewCompat;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.TextView;
 
 import org.thoughtcrime.securesms.components.RecyclerViewFastScroller;
-import org.thoughtcrime.securesms.components.StickyHeaderDecoration;
 import org.thoughtcrime.securesms.contacts.ContactSelectionListAdapter;
 import org.thoughtcrime.securesms.contacts.ContactSelectionListItem;
 import org.thoughtcrime.securesms.contacts.ContactsCursorLoader;
 import org.thoughtcrime.securesms.database.CursorRecyclerViewAdapter;
 import org.thoughtcrime.securesms.util.ViewUtil;
 
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -190,5 +193,182 @@ public class ContactSelectionListFragment extends    Fragment
   public interface OnContactSelectedListener {
     void onContactSelected(String number);
     void onContactDeselected(String number);
+  }
+
+  /**
+   * A sticky header decoration for android's RecyclerView.
+   */
+  public static class StickyHeaderDecoration extends RecyclerView.ItemDecoration {
+
+    private Map<Long, RecyclerView.ViewHolder> mHeaderCache;
+
+    private StickyHeaderAdapter mAdapter;
+
+    private boolean mRenderInline;
+
+    /**
+     * @param adapter the sticky header adapter to use
+     */
+    public StickyHeaderDecoration(StickyHeaderAdapter adapter, boolean renderInline) {
+      mAdapter = adapter;
+      mHeaderCache = new HashMap<>();
+      mRenderInline = renderInline;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void getItemOffsets(Rect outRect, View view, RecyclerView parent,
+                               RecyclerView.State state)
+    {
+      int position = parent.getChildAdapterPosition(view);
+
+      int headerHeight = 0;
+      if (position != RecyclerView.NO_POSITION && hasHeader(position)) {
+        View header = getHeader(parent, position).itemView;
+        headerHeight = getHeaderHeightForLayout(header);
+      }
+
+      outRect.set(0, headerHeight, 0, 0);
+    }
+
+    private boolean hasHeader(int position) {
+      if (position == 0) {
+        return true;
+      }
+
+      int previous = position - 1;
+      return mAdapter.getHeaderId(position) != mAdapter.getHeaderId(previous);
+    }
+
+    private RecyclerView.ViewHolder getHeader(RecyclerView parent, int position) {
+      final long key = mAdapter.getHeaderId(position);
+
+      if (mHeaderCache.containsKey(key)) {
+        return mHeaderCache.get(key);
+      } else {
+        final RecyclerView.ViewHolder holder = mAdapter.onCreateHeaderViewHolder(parent);
+        final View header = holder.itemView;
+
+        //noinspection unchecked
+        mAdapter.onBindHeaderViewHolder(holder, position);
+
+        int widthSpec = View.MeasureSpec.makeMeasureSpec(parent.getWidth(), View.MeasureSpec.EXACTLY);
+        int heightSpec = View.MeasureSpec.makeMeasureSpec(parent.getHeight(), View.MeasureSpec.UNSPECIFIED);
+
+        int childWidth = ViewGroup.getChildMeasureSpec(widthSpec,
+                                                       parent.getPaddingLeft() + parent.getPaddingRight(), header.getLayoutParams().width);
+        int childHeight = ViewGroup.getChildMeasureSpec(heightSpec,
+                                                        parent.getPaddingTop() + parent.getPaddingBottom(), header.getLayoutParams().height);
+
+        header.measure(childWidth, childHeight);
+        header.layout(0, 0, header.getMeasuredWidth(), header.getMeasuredHeight());
+
+        mHeaderCache.put(key, holder);
+
+        return holder;
+      }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void onDrawOver(Canvas c, RecyclerView parent, RecyclerView.State state) {
+      final int count = parent.getChildCount();
+
+      for (int layoutPos = 0; layoutPos < count; layoutPos++) {
+        final View child = parent.getChildAt(layoutPos);
+
+        final int adapterPos = parent.getChildAdapterPosition(child);
+
+        if (adapterPos != RecyclerView.NO_POSITION && (layoutPos == 0 || hasHeader(adapterPos))) {
+          View header = getHeader(parent, adapterPos).itemView;
+          c.save();
+          final int left = child.getLeft();
+          final int top = getHeaderTop(parent, child, header, adapterPos, layoutPos);
+          c.translate(left, top);
+          header.draw(c);
+          c.restore();
+        }
+      }
+    }
+
+    private int getHeaderTop(RecyclerView parent, View child, View header, int adapterPos,
+                             int layoutPos)
+    {
+      int headerHeight = getHeaderHeightForLayout(header);
+      int top = getChildY(parent, child) - headerHeight;
+      if (layoutPos == 0) {
+        final int count = parent.getChildCount();
+        final long currentId = mAdapter.getHeaderId(adapterPos);
+        // find next view with header and compute the offscreen push if needed
+        for (int i = 1; i < count; i++) {
+          int adapterPosHere = parent.getChildAdapterPosition(parent.getChildAt(i));
+          if (adapterPosHere != RecyclerView.NO_POSITION) {
+            long nextId = mAdapter.getHeaderId(adapterPosHere);
+            if (nextId != currentId) {
+              final View next = parent.getChildAt(i);
+              final int offset = getChildY(parent, next) - (headerHeight + getHeader(parent, adapterPosHere).itemView.getHeight());
+              if (offset < 0) {
+                return offset;
+              } else {
+                break;
+              }
+            }
+          }
+        }
+
+        top = Math.max(0, top);
+      }
+
+      return top;
+    }
+
+    private int getChildY(RecyclerView parent, View child) {
+      if (VERSION.SDK_INT < 11) {
+        Rect rect = new Rect();
+        parent.getChildVisibleRect(child, rect, null);
+        return rect.top;
+      } else {
+        return (int)ViewCompat.getY(child);
+      }
+    }
+
+    private int getHeaderHeightForLayout(View header) {
+      return mRenderInline ? 0 : header.getHeight();
+    }
+  }
+
+  /**
+   * The adapter to assist the {@link StickyHeaderDecoration} in creating and binding the header views.
+   *
+   * @param <T> the header view holder
+   */
+  public interface StickyHeaderAdapter<T extends RecyclerView.ViewHolder> {
+
+    /**
+     * Returns the header id for the item at the given position.
+     *
+     * @param position the item position
+     * @return the header id
+     */
+    long getHeaderId(int position);
+
+    /**
+     * Creates a new header ViewHolder.
+     *
+     * @param parent the header's view parent
+     * @return a view holder for the created view
+     */
+    T onCreateHeaderViewHolder(ViewGroup parent);
+
+    /**
+     * Updates the header view to reflect the header data for the given position
+     * @param viewHolder the header view holder
+     * @param position the header's item position
+     */
+    void onBindHeaderViewHolder(T viewHolder, int position);
   }
 }
